Fix shadowing of built-in list() function in CLI commands
The send command and batch operations were calling list() which was
being resolved to the Click command function instead of the built-in
list() function. This caused the CLI to execute the list command instead
of sending messages or performing batch operations.

Fixed by converting attachment/message ID arguments to tuples instead
of using list(), avoiding the name collision.